# specs/openapi.json
{
  "components" : {
    "schemas" : {
      "AbrechnungKurz" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "id" ],
        "type" : "object"
      },
...
      "ApiError" : {
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "message", "status" ]
      },
...
      "EinladungKurz" : {
        "properties" : {
          "anzahlPersonen" : {
            "format" : "int32",
            "type" : "integer"
          },
          "event" : {
            "$ref" : "#/components/schemas/EventResponse"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "partei" : {
            "$ref" : "#/components/schemas/ParteiKurz"
          },
          "status" : {
            "$ref" : "#/components/schemas/EinladungStatus"
          }
        },
        "required" : [ "event", "id", "partei", "status" ],
        "type" : "object"
      },
...
      "EinladungStatus" : {
        "enum" : [ "OFFEN", "ANGEMELDET", "ABGEMELDET" ],
        "type" : "string"
      },
...
      "EventResponse" : {
        "properties" : {
          "alternativerStandort" : {
            "type" : "string"
          },
          "datum" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "standort" : {
            "type" : "string"
          },
          "startzeit" : {
            "format" : "time-local",
            "type" : "string"
          },
          "zeitAufraumen" : {
            "format" : "time-local",
            "type" : "string"
          },
          "zeitAufstellen" : {
            "format" : "time-local",
            "type" : "string"
          }
        },
        "required" : [ "datum", "id", "standort", "startzeit" ],
        "type" : "object"
      },
      "KonsumationRequest" : {
        "properties" : {
          "anzahl" : {
            "format" : "int32",
            "type" : "integer"
          },
          "konsumationsangebotId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "teilnahmeId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "anzahl", "konsumationsangebotId", "teilnahmeId" ],
        "type" : "object"
      },
      "KonsumationResponse" : {
        "properties" : {
          "anzahl" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "konsumationsangebot" : {
            "$ref" : "#/components/schemas/KonsumationsangebotKurz"
          },
          "teilnahme" : {
            "$ref" : "#/components/schemas/TeilnahmeKurz"
          }
        },
        "required" : [ "anzahl", "id", "konsumationsangebot", "teilnahme" ],
        "type" : "object"
      },
      "KonsumationUpdateRequest" : {
        "properties" : {
          "anzahl" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "anzahl" ],
        "type" : "object"
      },
      "KonsumationsangebotKurz" : {
        "properties" : {
          "bezeichnung" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "preis" : {
            "type" : "number"
          }
        },
        "required" : [ "bezeichnung", "id", "preis" ],
        "type" : "object"
      },
...
      "ParteiKurz" : {
        "properties" : {
          "adresse" : {
            "type" : "string"
          },
          "bezeichnung" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "twintAktiv" : {
            "type" : "boolean"
          },
          "twintMobilenummer" : {
            "type" : "string"
          }
        },
        "required" : [ "adresse", "bezeichnung", "id", "twintAktiv" ],
        "type" : "object"
      },
...
      "TeilnahmeKurz" : {
        "properties" : {
          "einladung" : {
            "$ref" : "#/components/schemas/EinladungKurz"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "einladung", "id" ],
        "type" : "object"
      },
...
      "ZahlungRequest" : {
        "properties" : {
          "abrechnungId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "betrag" : {
            "type" : "number"
          },
          "datum" : {
            "format" : "date",
            "type" : "string"
          },
          "zahlungskanal" : {
            "$ref" : "#/components/schemas/Zahlungskanal"
          }
        },
        "required" : [ "abrechnungId", "betrag", "datum", "zahlungskanal" ],
        "type" : "object"
      },
      "ZahlungResponse" : {
        "properties" : {
          "abrechnung" : {
            "$ref" : "#/components/schemas/AbrechnungKurz"
          },
          "betrag" : {
            "type" : "number"
          },
          "datum" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "zahlungskanal" : {
            "$ref" : "#/components/schemas/Zahlungskanal"
          }
        },
        "required" : [ "abrechnung", "betrag", "datum", "id", "zahlungskanal" ],
        "type" : "object"
      },
      "Zahlungskanal" : {
        "enum" : [ "TWINT", "UEBERWEISUNG", "BAR" ],
        "type" : "string"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "REST-API des Quartierfest-Backends. Authentifizierung via POST /api/auth/login (HS256-JWT, AUTH-002).",
    "title" : "Quartierfest API",
    "version" : "0.0.1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/konsumationen" : {
      "get" : {
        "operationId" : "findAll_6",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/KonsumationResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Fehler im einheitlichen Format (ERROR-001)"
          }
        },
        "tags" : [ "konsumation-controller" ]
      },
      "post" : {
        "operationId" : "create_6",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KonsumationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/KonsumationResponse"
                }
              }
            },
            "description" : "OK"
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Fehler im einheitlichen Format (ERROR-001)"
          }
        },
        "tags" : [ "konsumation-controller" ]
      }
    },
    "/api/konsumationen/{id}" : {
      "delete" : {
        "operationId" : "delete_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Fehler im einheitlichen Format (ERROR-001)"
          }
        },
        "tags" : [ "konsumation-controller" ]
      },
      "put" : {
        "operationId" : "update_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KonsumationUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/KonsumationResponse"
                }
              }
            },
            "description" : "OK"
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Fehler im einheitlichen Format (ERROR-001)"
          }
        },
        "tags" : [ "konsumation-controller" ]
      }
    },
...
    "/api/zahlungen" : {
      "get" : {
        "operationId" : "findAll",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ZahlungResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Fehler im einheitlichen Format (ERROR-001)"
          }
        },
        "tags" : [ "zahlung-controller" ]
      },
      "post" : {
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ZahlungRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ZahlungResponse"
                }
              }
            },
            "description" : "OK"
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Fehler im einheitlichen Format (ERROR-001)"
          }
        },
        "tags" : [ "zahlung-controller" ]
      }
    },
    "/api/zahlungen/{id}" : {
      "delete" : {
        "operationId" : "delete_9",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Fehler im einheitlichen Format (ERROR-001)"
          }
        },
        "tags" : [ "zahlung-controller" ]
      }
    }
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost:8080"
  } ]
}